requirements.txt
fastapi
uvicorn
pydantic
dspy-ai
python-dotenv
requests
pandas
numpy
alpaca-py
psycopg2-binary
ollama
yfinance
